// app/interfaces/globalStatsDTO.ts
export interface GlobalStatsDTO {
    totalProjectsInProgress: number;
    totalProjectsPending: number;
    totalProjectsCompleted: number;
    
    totalTasksInProgress: number;
    totalTasksPending: number;
    totalTasksCompleted: number;
    totalUsers: number;
    totalProjects: number;
    totalTasks: number;
    totalDepartments: number;
    totalWomen: number;
    totalMen: number;
}